pip install with --no-cache-dir
for reasons only clear to the python packaging gods, pip
first builds wheels in a cache, and then install them. but
if two packages we are installing depend on one another,
when the second package (which depends on a newer version of
the first) is bdist_wheel'd the build will fail. however when
pip then subsequently executes setup.py install, the second build
will succeed as the package which is depended on is now
present.

just disable this madness by disabling the wheel cache and
rely on ordering to express dependencies.

may god have mercy on your soul